Split up "event log" test
Into a test for the creation and for the page.
